Update Ponto and Caminho ids to int and add TipoPonto to Ponto

+++ src/main/java/com/milleddy/movucsal/repository/CaminhoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CaminhoRepository extends JpaRepository<Caminho, Long> {
+public interface CaminhoRepository extends JpaRepository<Caminho, Integer> {
 
 }
