+++ src/main/java/com/milleddy/movucsal/repository/PontoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PontoRepository extends JpaRepository<Ponto, Long> {
+public interface PontoRepository extends JpaRepository<Ponto, Integer> {
 }
